+++ tools/pythonpkg/src/include/duckdb_python/numpy/raw_array_wrapper.hpp
@@ -9,25 +9,25 @@
 #pragma once
 
 #include "duckdb_python/pybind11/pybind_wrapper.hpp"
-#include "duckdb_python/numpy/raw_array_wrapper.hpp"
 #include "duckdb.hpp"
 
 namespace duckdb {
 
-struct ArrayWrapper {
-	explicit ArrayWrapper(const LogicalType &type, const ClientProperties &client_properties, bool pandas = false);
+struct RawArrayWrapper {
 
-	unique_ptr<RawArrayWrapper> data;
-	unique_ptr<RawArrayWrapper> mask;
-	bool requires_mask;
-	const ClientProperties client_properties;
-	bool pandas;
+	explicit RawArrayWrapper(const LogicalType &type);
+
+	py::array array;
+	data_ptr_t data;
+	LogicalType type;
+	idx_t type_width;
+	idx_t count;
 
 public:
+	static string DuckDBToNumpyDtype(const LogicalType &type);
 	void Initialize(idx_t capacity);
 	void Resize(idx_t new_capacity);
 	void Append(idx_t current_offset, Vector &input, idx_t count);
-	py::object ToArray(idx_t count) const;
 };
 
 } // namespace duckdb
