+++ tools/pythonpkg/src/include/duckdb_python/pandas/pandas_bind.hpp
@@ -8,8 +8,13 @@
 
 namespace duckdb {
 
-struct RegisteredArray;
 class ClientContext;
+
+struct RegisteredArray {
+	explicit RegisteredArray(py::array numpy_array) : numpy_array(std::move(numpy_array)) {
+	}
+	py::array numpy_array;
+};
 
 struct PandasColumnBindData {
 	NumpyType numpy_type;
