+++ tools/pythonpkg/src/pandas/bind.cpp
@@ -1,5 +1,4 @@
 #include "duckdb_python/pandas/pandas_bind.hpp"
-#include "duckdb_python/numpy/array_wrapper.hpp"
 #include "duckdb_python/pandas/pandas_analyzer.hpp"
 #include "duckdb_python/pandas/column/pandas_numpy_column.hpp"
