+++ tools/pythonpkg/src/pyresult.cpp
@@ -163,7 +163,8 @@ void InsertCategory(QueryResult &result, unordered_map<idx_t, py::list> &categor
 	}
 }
 
-py::dict DuckDBPyResult::FetchNumpyInternal(bool stream, idx_t vectors_per_chunk) {
+py::dict DuckDBPyResult::FetchNumpyInternal(bool stream, idx_t vectors_per_chunk,
+                                            unique_ptr<NumpyResultConversion> conversion_p) {
 	if (!result) {
 		throw InvalidInputException("result closed");
 	}
@@ -176,7 +177,11 @@ py::dict DuckDBPyResult::FetchNumpyInternal(bool stream, idx_t vectors_per_chunk
 		initial_capacity = materialized.RowCount();
 	}
 
-	NumpyResultConversion conversion(result->types, initial_capacity, result->client_properties);
+	if (!conversion_p) {
+		conversion_p = make_uniq<NumpyResultConversion>(result->types, initial_capacity, result->client_properties);
+	}
+	auto &conversion = *conversion_p;
+
 	if (result->type == QueryResultType::MATERIALIZED_RESULT) {
 		auto &materialized = result->Cast<MaterializedQueryResult>();
 		for (auto &chunk : materialized.Collection().Chunks()) {
@@ -259,6 +264,7 @@ PandasDataFrame DuckDBPyResult::FrameFromNumpy(bool date_as_object, const py::ha
 }
 
 PandasDataFrame DuckDBPyResult::FetchDF(bool date_as_object) {
+
 	return FrameFromNumpy(date_as_object, FetchNumpyInternal());
 }
